server/requirements.txt
asgiref==3.6.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
cloudinary==1.38.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.1
dnspython==2.5.0
email-validator==2.1.0.post1
filelock==3.13.1
Flask==2.2.5
flask-mongoengine==0.9.5
Flask-WTF==1.2.1
fsspec==2024.2.0
ftfy==6.1.3
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
install==1.3.5
instaloader==4.10.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.27.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pillow==10.2.0
proglog==0.1.10
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
